==> 3_IMDB_vectorization_w2v.ipynb <==
# IMDB Data Vectorization

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split 
import gensim

# df = pd.read_csv('cleaned_stemmed_IMDb_reviews.csv')
df = pd.read_csv('cleaned_non_stemmed_IMDb_reviews.csv')
df.head()

#### 1. Tokenization of our reviews

# tokenizing reviews
tokenized_review = df['review'].apply(lambda x: x.split())

#### 3. Training a Word2Vec model

# SETTING MODEL PARAMETERS
model_w2v = gensim.models.Word2Vec(
    tokenized_review,
    vector_size = 1000, # desired number of features
    window = 5, # context window
    min_count = 2, # ignore all words with total frequency lower than 2
    sg = 1,
    hs = 0,
    negative = 10,
    workers = 10,
    seed = 34
)

# TRAINING MODEL ON CORPUS
model_w2v.train(
    tokenized_review,
    total_examples = len(df['review']),
    epochs = 20
)

#### 4. Transforming our tokenized features into vectors 

# FUNC: CREATES VECTOR FOR EACH REVIEW BY TAKING THE AVERAGE OF THE VECTORS OF WORDS PRESENT IN REVIEW
def word_vector(tokens, size):
    vec = np.zeros(size).reshape((1, size))
    count = 0
    for word in tokens:
        try:
            vec += model_w2v.wv[word].reshape((1, size))
            count += 1
        except KeyError:    # case where token is not in vocab
            continue
    if count !=0:
        vec /= count
    return vec

# applying above method to each review
wordvec_arrays = np.zeros((len(tokenized_review), 1000))
for i in range(len(tokenized_review)):
    wordvec_arrays[i,:] = word_vector(tokenized_review[i], 1000)

wordvec_df = pd.DataFrame(wordvec_arrays)
wordvec_df.shape

#### 5. Rejoining our vectors alongside our target feature and saving to .CSV file to be used later for modeling development

wordvec_df['sentiment'] = df['sentiment']
wordvec_df.shape

wordvec_df.to_csv('IMDb_nonstemmed_w2v_1000v_data.csv', index=False)
# wordvec_df.to_csv('IMDb_nonstemmed_w2v_500v_data.csv', index=False)
# wordvec_df.to_csv('IMDb_stemmed_w2v_data.csv', index=False)
# wordvec_df.to_csv('IMDb_nonstemmed_vectorized_data.csv', index=False)
