==> 5_IMDB_LR_SVM_XGBoost_nonstemmed_500v_Tunned.ipynb <==
# IMDB Model Development, Tuning, and Evaluation

Goal: Achieve highest accuracy possible in using classification to distinguish between positive and negative textual reviews.

Models to be developed:
1. Logistic Regression
2. SVM
3. XGBoost


import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.svm import LinearSVC
from sklearn.linear_model import LogisticRegression
from xgboost import XGBClassifier

##### Reading in Data
Textual data has already been vectorized using Word2Vec resulting in 500 features representing each review and 1 target feature, Sentiment.

df = pd.read_csv('IMDb_nonstemmed_w2v_500v_data.csv')
print(df.shape)
df.head()

#### Splitting into Training and Testing sets
We will be splitting the data into 80% training and 20% testing sets.

The __training set__ will be used for __fitting__.

The __testing set__ will be used for __final model evaluations__.

# all features except target
X = df.drop('sentiment', axis=1)    # to be used for 
# only target feature (sentiment)
y = df['sentiment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.20, random_state=34)

#### __Note: For the following models we will be using the optimal hyperparameters found when training models using a vector size of 200__

### 1. Logistic Regression

lr_model = LogisticRegression(C = 1.0, max_iter=1000, penalty='l1', solver = 'saga')
lr_model.fit(X_train, y_train)
y_pred = lr_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

### 2. SVM


##### 2.1 Initial Modeling

svc_model = LinearSVC(C = 10, loss = 'squared_hinge', max_iter = 1000, penalty= 'l2')
svc_model.fit(X_train, y_train)
y_pred = svc_model.predict(X_test)
print(classification_report(y_test,y_pred, digits=4))

### 3. XGBoost

xgb_model = XGBClassifier(learning_rate = 0.2, max_depth = 9, ns_estimator = 100)
xgb_model.fit(X_train, y_train)

y_pred = xgb_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))
