==> 6_IMDb_RF_DT_MLP_500v_Tuned.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split, GridSearchCV, KFold
from sklearn.neural_network import MLPClassifier
from sklearn.metrics import classification_report
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier

df = pd.read_csv('IMDb_nonstemmed_w2v_500v_data.csv')
df.shape


X = df.drop('sentiment', axis = 1)
y = df['sentiment']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=43)
# ensuring we have the same splits for all models
kf = KFold(n_splits=5)

### 1. Random Forest

1.1 Initial Modeling 

rf_model = RandomForestClassifier()
rf_model.fit(X_train, y_train)
y_pred = rf_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

1.2 Random Forest tuning using GridSearchCV

# hyper parameters to be tuned
params = [{
    'n_estimators': [50, 100, 200],
    'criterion': ['gini', 'entropy']
}]

rf_GCV = GridSearchCV(RandomForestClassifier(), param_grid=params, cv = kf, verbose=4)
best_rf = rf_GCV.fit(X_train, y_train)

Optimal hyperparameters:

print(best_rf.best_params_)

#79 mins to fi
best_rf.best_estimator_

y_pred = best_rf.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

### 2. Decision Tree

2.1 Inital Modeling

dt_model = DecisionTreeClassifier()
dt_model.fit(X_train, y_train)
y_pred = dt_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

2.2 Decision Tree tuning using GridSearchCV

# hyper parameters to be tuned
params = [{
    'criterion': ['gini', 'entropy'],
    'max_depth': ['None', 50, 100]
}]
dt_GCV = GridSearchCV(DecisionTreeClassifier(), param_grid=params, cv = kf, verbose=4)
best_dt = dt_GCV.fit(X_train, y_train)

best_dt.best_params_

# fit took 19 minutes
best_dt.best_estimator_

y_pred = best_dt.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

#### Multi-Layer Perceptron

3.1 Initial Modeling

mlpModel = MLPClassifier()
mlpModel.fit(X_train, y_train)

y_pred = mlpModel.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

3.2 MLPClassifer tuning with GridSearchCV

params = [{
    'hidden_layer_sizes': [(100,), (200,)],
    'activation': ['logistic', 'adam']
}]

mlp_GCV = GridSearchCV(MLPClassifier(), param_grid=params, cv = kf, verbose=4)
best_mlp = mlp_GCV.fit(X_train, y_train)

best_mlp.best_params_

# fit took 19 minutes
best_mlp.best_estimator_

y_pred = best_mlp.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))
