==> 7_IMDb_modeling_with_varying_#_vectors.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.linear_model import LogisticRegression

df = pd.read_csv('IMDb_nonstemmed_w2v_300v_data.csv')
df.shape

X = df.drop('sentiment', axis=1)
y = df['sentiment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.20, random_state=34)

# LR performance before hyperparameter tuning (default params)
lr_model = LogisticRegression()
lr_model.fit(X_train, y_train)
y_pred = lr_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

df_500 = pd.read_csv('IMDb_nonstemmed_w2v_500v_data.csv')
df_500.shape

X = df_500.drop('sentiment', axis=1)
y = df_500['sentiment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.20, random_state=34)

# LR performance before hyperparameter tuning (default params)
lr_model = LogisticRegression()
lr_model.fit(X_train, y_train)
y_pred = lr_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))

df_1000 = pd.read_csv('IMDb_nonstemmed_w2v_1000v_data.csv')
df_1000.shape

X = df_1000.drop('sentiment', axis=1)
y = df_1000['sentiment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.20, random_state=34)

# LR performance before hyperparameter tuning (default params)
lr_model = LogisticRegression()
lr_model.fit(X_train, y_train)
y_pred = lr_model.predict(X_test)
print(classification_report(y_test, y_pred, digits=4))
